move tests back for linting
